Refactor addons init to addons manager
